requirements.txt
streamlit
numpy
scipy.stats